llm_learning/notebook/requirements.txt
zhipuai
chromadb
pymupdf
langchain-community
unstructured